feat: add approach entries
